# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restaurante API",
    "version": "1.0.0",
    "description": "API para gerenciamento de pedidos e cardápio do restaurante."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/pedidos": {
      "get": {
        "summary": "Lista todos os pedidos",
        "responses": {
          "200": {
            "description": "Lista de pedidos"
          }
        }
      },
      "post": {
        "summary": "Cria um novo pedido",
        "description": "Cria um pedido apenas se o prato existir no cardápio do banco de dados.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prato": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido criado"
          },
          "400": {
            "description": "Erro de validação ou prato não existe no cardápio"
          }
        }
      }
    },
    "/menu": {
      "get": {
        "summary": "Lista o cardápio do banco de dados",
        "responses": {
          "200": {
            "description": "Lista de itens do cardápio cadastrados no banco"
          }
        }
      }
    },
    "/pedidos/{id}": {
      "get": {
        "summary": "Busca um pedido pelo ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      },
      "put": {
        "summary": "Atualiza o status de um pedido",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido atualizado"
          },
          "400": {
            "description": "Erro de validação"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Exclui um pedido",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pedido excluído"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      }
    },
    "/pedidos/status/{status}": {
      "get": {
        "summary": "Filtra pedidos por status",
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos filtrados"
          },
          "404": {
            "description": "Nenhum pedido encontrado"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}